requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
numpy==2.2.1
scipy==1.15.0
pytest==8.3.4
pytest-asyncio==0.25.0
